requirements.txt
django-taggit
django-allauth
django_threadedcomments
south